On the authzperf branch: Change the type of the sequence number in an ACL from apr_int64_t to just plain int:   * ACLs are stored in an array, and apr_array_t::nelts is an int;   * we don't need an exact-sized type for the sequence number;   * anyone with more than 2^31 access rules, or who tries to compile     Subversion on a 16-bit platform, is going to be in trouble anyway.
* subversion/libsvn_repos/authz.h
  (authz_acl_t): Change the type of sequence_number to int.
* subversion/tests/libsvn_repos/authz-test.c
  (test_authz_parse_tng): Change the format and remove the cast
   for printing the sequence number.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/authzperf@1616083 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/tests/libsvn_repos/authz-test.c
@@ -119,13 +119,13 @@ test_authz_parse_tng(const svn_test_opts_t *opts,
         svn_authz__acl_get_access(&access, acl, check_user, check_repo);
       int j;
 
-      printf("%s%s%s   Sequence:   %ld\n"
+      printf("%s%s%s   Sequence:   %d\n"
              "   Repository: [%s]\n"
              "   Rule:  %s[/%s]\n",
              (has_access ? "Match = " : ""),
              (has_access ? access_string(access) : ""),
              (has_access ? "\n" : ""),
-             (long)acl->sequence_number,
+             acl->sequence_number,
              acl->repos,
              (acl->glob ? "glob:" : "     "),
              acl->rule);
